+++ app/src/main/java/com/dds/theecogame/domain/factory/Validator.kt
@@ -3,4 +3,5 @@ package com.dds.theecogame.domain.factory
 interface Validator {
     fun validate(input: String): Boolean
     fun getError(): String
+    fun setPasswordValidation(type: Boolean)
 }
