+++ app/src/main/java/com/dds/theecogame/domain/factory/validators/EmailValidator.kt
@@ -46,4 +46,7 @@ class EmailValidator : Validator {
 
         return true
     }
+
+    override fun setPasswordValidation(type: Boolean) {
+    }
 }
