+++ app/src/main/java/com/dds/theecogame/domain/factory/validators/PasswordValidator.kt
@@ -4,39 +4,33 @@ import com.dds.theecogame.R
 import com.dds.theecogame.data.repository.UserRepositoryImpl
 import com.dds.theecogame.domain.factory.Validator
 import com.dds.theecogame.domain.repository.UserRepository
+import com.dds.theecogame.domain.strategy.passwordValidators.StrongPasswordValidation
+import com.dds.theecogame.domain.strategy.passwordValidators.WeakPasswordValidation
 
 class PasswordValidator : Validator {
     private var error: String = String()
+    private var validationType: Boolean = false
 
     override fun getError(): String {
         return this.error
     }
 
     override fun validate(password: String): Boolean {
-        //check principal
-        val lowercase = Regex(".*[a-z].*")
-        val uppercase = Regex(".*[A-Z].*")
-        val number = Regex(".*[\\d].*")
+        if (validationType) {
+            val strongValidationType = StrongPasswordValidation()
+            val result = strongValidationType.validatePassword(password)
+            error = strongValidationType.getError()
+            return result
 
-        if (password.isEmpty()) return true
-        if (password.length < 8) {
-            error = "Debe contener al menos 8 caracteres"
-            return false
+        } else {
+            val weakValidationType = WeakPasswordValidation()
+            val result = weakValidationType.validatePassword(password)
+            error = weakValidationType.getError()
+            return result
         }
-        if (!lowercase.matches(password)) {
-            error = "Debe contener al menos una minúscula"
-            return false
-        }
-        if (!uppercase.matches(password)) {
-            error = "Debe contener al menos una mayúscula"
-            return false
-        }
-        if (!number.matches(password)) {
-            error = "Debe contener al menos un número"
-            return false
-        }
-        //check con llamada a api NO es necesaria
+    }
 
-        return true
+    override fun setPasswordValidation(type: Boolean) {
+        validationType = type
     }
 }
