+++ app/src/main/java/com/dds/theecogame/domain/strategy/PasswordValidationType.kt
@@ -1,5 +1,6 @@
 package com.dds.theecogame.domain.strategy
 
 interface PasswordValidationType {
-    fun validatePassword(): Boolean
+    fun validatePassword(password: String): Boolean
+    fun getError(): String
 }
