+++ app/src/main/java/com/dds/theecogame/domain/strategy/passwordValidators/WeakPasswordValidation.kt
@@ -3,7 +3,20 @@ package com.dds.theecogame.domain.strategy.passwordValidators
 import com.dds.theecogame.domain.strategy.PasswordValidationType
 
 class WeakPasswordValidation : PasswordValidationType {
-    override fun validatePassword(): Boolean {
-        TODO("Not yet implemented")
+    private var error: String = String()
+
+    override fun validatePassword(password: String): Boolean {
+
+        if (password.isEmpty()) return true
+        if (password.length < 8) {
+            error = "Debe contener al menos 8 caracteres"
+            return false
+        }
+
+        return true
+    }
+
+    override fun getError(): String {
+        return this.error
     }
 }
